refactor: Like 도메인을 Post 도메인 안에 포함.
