+++ src/main/java/org/ahpuh/surf/user/domain/User.java
@@ -4,8 +4,8 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.ahpuh.surf.category.domain.Category;
 import org.ahpuh.surf.common.entity.BaseEntity;
-import org.ahpuh.surf.like.domain.Like;
 import org.ahpuh.surf.post.domain.Post;
+import org.ahpuh.surf.post.domain.like.Like;
 import org.ahpuh.surf.user.domain.follow.Follow;
 import org.ahpuh.surf.user.dto.request.UserUpdateRequestDto;
 import org.hibernate.annotations.SQLDelete;
